Update main.py
I implemented a for loop that collects the duration of two destinations in a dictionary (Line: 38-48). Every time those two destinations get called, the duration just gets returned. This makes my program way quicker and the API does not have to be called as much.
 
To even increase the speed, I implemented another variable to store the reversed duration of two destinations as well in the dictionary cache. This will allow the program to call the duration, without having to run the code twice.